modified build compilation stuff
