+++ app/api/articles/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse, NextRequest } from "next/server";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/lib/auth";
 import { getArticlesByWriterId, createArticle } from "@/app/lib/database/query";
 import { Session } from "next-auth";
