+++ app/api/user/change-password/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/lib/auth";
 import { getUserByEmail, updateUserPassword } from "@/app/lib/database/query";
 import { compare } from "bcrypt";
