+++ app/api/user/me/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/lib/auth";
 import { getUserByEmail } from "@/app/lib/database/query";
 
 export async function GET() {
