+++ app/api/user/update-profile/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/lib/auth";
 import { updateUserProfile } from "@/app/lib/database/query";
 
 export async function PUT(request: NextRequest) {
