+++ app/lib/auth.ts
@@ -1,0 +1,60 @@
+import CredentialsProvider from "next-auth/providers/credentials";
+import { compare } from "bcrypt";
+import { getUserByEmail } from "@/app/lib/database/query";
+
+export const authOptions = {
+  providers: [
+    CredentialsProvider({
+      name: "credentials",
+      credentials: {
+        email: { label: "Email", type: "email" },
+        password: { label: "Password", type: "password" },
+      },
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      async authorize(credentials: any) {
+        if (!credentials?.email || !credentials?.password) {
+          return null;
+        }
+
+        const user = await getUserByEmail(credentials.email);
+        if (!user) return null;
+
+        const isValid = await compare(credentials.password, user.password);
+        if (!isValid) return null;
+
+        return {
+          id: user.id,
+          email: user.email,
+          name: user.first_name || user.email,
+          auth_level: user.auth_level,
+        };
+      },
+    }),
+  ],
+  session: { strategy: "jwt" as const },
+  callbacks: {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    async jwt({ token, user }: any) {
+      if (user) {
+        token.id = user.id;
+        token.auth_level = user.auth_level;
+      }
+      return token;
+    },
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    async session({ session, token }: any) {
+      if (token && session.user) {
+        session.user.id = token.id as string;
+        session.user.auth_level = token.auth_level as
+          | "admin"
+          | "team"
+          | "basic";
+      }
+      return session;
+    },
+  },
+  pages: {
+    signIn: "/auth/signin",
+  },
+  secret: process.env.NEXTAUTH_SECRET,
+};
